pyproject.toml
[tool.poetry]
name = "talkdoc_core"
version = "0.1.0"
description = ""
authors = ["dkr <deepukreddy123@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10, <3.13"
pypdf = "^4.2.0"
openai = "^1.23.6"
python-dotenv = "^1.0.1"
streamlit = "^1.33.0"
streamlit-chat = "^0.1.1"
ipykernel = "^6.29.4"
streamlit-pdf-viewer = "^0.0.11"
streamlit-mic-recorder = "^0.0.8"
pypdftk = "^0.5"
rich = "^13.9.4"
pymupdf = "^1.25.1"
streamlit-authenticator = "^0.4.1"
agno = "^1.1.2"
duckduckgo-search = "^7.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
